outsourced VisSetting class
